Moved attributes that are used for constructing url to separate list. Added list for attributes needed for delete-operation.

+++ gitlab.py
@@ -246,7 +246,8 @@ class Gitlab(object):
 
     def list(self, obj_class, **kwargs):
         missing = []
-        for k in obj_class.requiredListAttrs:
+        for k in chain(obj_class.requiredUrlAttrs,
+                       obj_class.requiredListAttrs):
             if k not in kwargs:
                 missing.append(k)
         if missing:
@@ -255,10 +256,16 @@ class Gitlab(object):
 
         url = self.constructUrl(id_=None, obj=obj_class, parameters=kwargs)
         headers = self.createHeaders()
+
+        # Remove attributes that are used in url so that there is only
+        # url-parameters left
+        params = kwargs.copy()
+        for attribute in obj_class.requiredUrlAttrs:
+            del params[attribute]
+
         try:
-            r = requests.get(url, params=kwargs, headers=headers,
-                             verify=self.ssl_verify,
-                             timeout=self.timeout)
+            r = requests.get(url, params=params, headers=headers,
+                             verify=self.ssl_verify, timeout=self.timeout)
         except:
             raise GitlabConnectionError(
                 "Can't connect to GitLab server (%s)" % self._url)
@@ -280,7 +287,8 @@ class Gitlab(object):
 
     def get(self, obj_class, id=None, **kwargs):
         missing = []
-        for k in obj_class.requiredGetAttrs:
+        for k in chain(obj_class.requiredUrlAttrs,
+                       obj_class.requiredGetAttrs):
             if k not in kwargs:
                 missing.append(k)
         if missing:
@@ -289,8 +297,15 @@ class Gitlab(object):
 
         url = self.constructUrl(id_=id, obj=obj_class, parameters=kwargs)
         headers = self.createHeaders()
+
+        # Remove attributes that are used in url so that there is only
+        # url-parameters left
+        params = kwargs.copy()
+        for attribute in obj_class.requiredUrlAttrs:
+            del params[attribute]
+
         try:
-            r = requests.get(url, params=kwargs,
+            r = requests.get(url, params=params,
                              headers=headers,
                              verify=self.ssl_verify,
                              timeout=self.timeout)
@@ -304,11 +319,25 @@ class Gitlab(object):
             self._raiseErrorFromResponse(r, GitlabGetError)
 
     def delete(self, obj, **kwargs):
-        url = self.constructUrl(id_=obj.id, obj=obj, parameters=obj.__dict__)
+        params = obj.__dict__.copy()
+        params.update(kwargs)
+        missing = []
+        for k in chain(obj.requiredUrlAttrs, obj.requiredDeleteAttrs):
+            if k not in params:
+                missing.append(k)
+        if missing:
+            raise GitlabDeleteError('Missing attribute(s): %s' %
+                                    ", ".join(missing))
+
+        url = self.constructUrl(id_=obj.id, obj=obj, parameters=params)
         headers = self.createHeaders()
+
+        for attribute in obj.requiredUrlAttrs:
+            del params[attribute]
+
         try:
             r = requests.delete(url,
-                                params=kwargs,
+                                params=params,
                                 headers=headers,
                                 verify=self.ssl_verify,
                                 timeout=self.timeout)
@@ -325,7 +354,7 @@ class Gitlab(object):
         params = obj.__dict__.copy()
         params.update(kwargs)
         missing = []
-        for k in obj.requiredCreateAttrs:
+        for k in chain(obj.requiredUrlAttrs, obj.requiredCreateAttrs):
             if k not in params:
                 missing.append(k)
         if missing:
@@ -355,7 +384,7 @@ class Gitlab(object):
         params = obj.__dict__.copy()
         params.update(kwargs)
         missing = []
-        for k in obj.requiredCreateAttrs:
+        for k in chain(obj.requiredUrlAttrs, obj.requiredCreateAttrs):
             if k not in params:
                 missing.append(k)
         if missing:
@@ -515,8 +544,10 @@ class GitlabObject(object):
     canCreate = True
     canUpdate = True
     canDelete = True
+    requiredUrlAttrs = []
     requiredListAttrs = []
     requiredGetAttrs = []
+    requiredDeleteAttrs = []
     requiredCreateAttrs = []
     optionalCreateAttrs = []
     idAttr = 'id'
@@ -679,19 +710,19 @@ class GitlabObject(object):
 class UserKey(GitlabObject):
     _url = '/users/%(user_id)s/keys'
     canGet = False
-    canList = True
     canUpdate = False
-    canDelete = True
-    requiredCreateAttrs = ['user_id', 'title', 'key']
+    requiredUrlAttrs = ['user_id']
+    requiredCreateAttrs = ['title', 'key']
 
 
 class User(GitlabObject):
     _url = '/users'
     shortPrintAttr = 'username'
-    requiredCreateAttrs = ['email', 'password', 'username', 'name']
-    optionalCreateAttrs = ['skype', 'linkedin', 'twitter', 'projects_limit',
-                           'extern_uid', 'provider', 'bio', 'admin',
-                           'can_create_group']
+    # FIXME: password is required for create but not for update
+    requiredCreateAttrs = ['email', 'username', 'name']
+    optionalCreateAttrs = ['password', 'skype', 'linkedin', 'twitter',
+                           'projects_limit', 'extern_uid', 'provider',
+                           'bio', 'admin', 'can_create_group', 'website_url']
 
     def Key(self, id=None, **kwargs):
         return UserKey._getListOrObject(self.gitlab, id,
@@ -721,13 +752,14 @@ class GroupMember(GitlabObject):
     _url = '/groups/%(group_id)s/members'
     canGet = False
     canUpdate = False
-    requiredCreateAttrs = ['group_id', 'user_id', 'access_level']
-    requiredDeleteAttrs = ['group_id', 'user_id']
+    requiredUrlAttrs = ['group_id']
+    requiredCreateAttrs = ['access_level']
     shortPrintAttr = 'username'
 
 
 class Group(GitlabObject):
     _url = '/groups'
+    canUpdate = False
     _constructorTypes = {'projects': 'Project'}
     requiredCreateAttrs = ['name', 'path']
     shortPrintAttr = 'name'
@@ -770,12 +802,11 @@ class Issue(GitlabObject):
 
 class ProjectBranch(GitlabObject):
     _url = '/projects/%(project_id)s/repository/branches'
+    _constructorTypes = {'author': 'User', "committer": "User"}
     idAttr = 'name'
     canUpdate = False
-    requiredGetAttrs = ['project_id']
-    requiredListAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'branch_name', 'ref']
-    requiredDeleteAttrs = ['project_id']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['branch_name', 'ref']
     _constructorTypes = {'commit': 'ProjectCommit'}
 
     def protect(self, protect=True, **kwargs):
@@ -804,7 +835,7 @@ class ProjectCommit(GitlabObject):
     canDelete = False
     canUpdate = False
     canCreate = False
-    requiredListAttrs = ['project_id']
+    requiredUrlAttrs = ['project_id']
     shortPrintAttr = 'title'
 
     def diff(self, **kwargs):
@@ -830,9 +861,8 @@ class ProjectCommit(GitlabObject):
 class ProjectKey(GitlabObject):
     _url = '/projects/%(project_id)s/keys'
     canUpdate = False
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'title', 'key']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['title', 'key']
 
 
 class ProjectEvent(GitlabObject):
@@ -841,16 +871,19 @@ class ProjectEvent(GitlabObject):
     canDelete = False
     canUpdate = False
     canCreate = False
-    requiredListAttrs = ['project_id']
+    requiredUrlAttrs = ['project_id']
     shortPrintAttr = 'target_title'
 
 
 class ProjectHook(GitlabObject):
     _url = '/projects/%(project_id)s/hooks'
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'url']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['url']
+    optionalCreateAttrs = ['push_events', 'issues_events',
+                           'merge_requests_events', 'tag_push_events']
+
     shortPrintAttr = 'url'
+
 
 
 class ProjectIssueNote(GitlabObject):
@@ -858,9 +891,8 @@ class ProjectIssueNote(GitlabObject):
     _constructorTypes = {'author': 'User'}
     canUpdate = False
     canDelete = False
-    requiredListAttrs = ['project_id', 'issue_id']
-    requiredGetAttrs = ['project_id', 'issue_id']
-    requiredCreateAttrs = ['project_id', 'body']
+    requiredUrlAttrs = ['project_id', 'issue_id']
+    requiredCreateAttrs = ['body']
 
 
 class ProjectIssue(GitlabObject):
@@ -868,12 +900,12 @@ class ProjectIssue(GitlabObject):
     _constructorTypes = {'author': 'User', 'assignee': 'User',
                          'milestone': 'ProjectMilestone'}
     canDelete = False
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'title']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['title']
     optionalCreateAttrs = ['description', 'assignee_id', 'milestone_id',
                            'labels']
     shortPrintAttr = 'title'
+
 
     def Note(self, id=None, **kwargs):
         return ProjectIssueNote._getListOrObject(self.gitlab, id,
@@ -884,9 +916,8 @@ class ProjectIssue(GitlabObject):
 
 class ProjectMember(GitlabObject):
     _url = '/projects/%(project_id)s/members'
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'user_id', 'access_level']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['access_level']
     shortPrintAttr = 'username'
 
 
@@ -895,9 +926,8 @@ class ProjectNote(GitlabObject):
     _constructorTypes = {'author': 'User'}
     canUpdate = False
     canDelete = False
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'body']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['body']
 
 
 class ProjectTag(GitlabObject):
@@ -906,30 +936,29 @@ class ProjectTag(GitlabObject):
     canGet = False
     canDelete = False
     canUpdate = False
-    canCreate = False
-    requiredListAttrs = ['project_id']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['tag_name', 'ref']
+    optionalCreateattrs = ['message']
     shortPrintAttr = 'name'
 
 
 class ProjectMergeRequestNote(GitlabObject):
     _url = '/projects/%(project_id)s/merge_requests/%(merge_request_id)s/notes'
     _constructorTypes = {'author': 'User'}
-    canGet = False
-    canCreate = False
     canUpdate = False
     canDelete = False
-    requiredListAttrs = ['project_id', 'merge_request_id']
+    requiredUrlAttrs = ['project_id', 'merge_request_id']
+    requiredCreateAttrs = ['body']
 
 
 class ProjectMergeRequest(GitlabObject):
     _url = '/projects/%(project_id)s/merge_requests'
     _constructorTypes = {'author': 'User', 'assignee': 'User'}
     canDelete = False
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'source_branch',
-                           'target_branch', 'title']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['source_branch', 'target_branch', 'title']
     optionalCreateAttrs = ['assignee_id']
+
 
     def Note(self, id=None, **kwargs):
         return ProjectMergeRequestNote._getListOrObject(
@@ -942,9 +971,8 @@ class ProjectMergeRequest(GitlabObject):
 class ProjectMilestone(GitlabObject):
     _url = '/projects/%(project_id)s/milestones'
     canDelete = False
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'title']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['title']
     optionalCreateAttrs = ['description', 'due_date', 'state_event']
     shortPrintAttr = 'title'
 
@@ -954,19 +982,18 @@ class ProjectSnippetNote(GitlabObject):
     _constructorTypes = {'author': 'User'}
     canUpdate = False
     canDelete = False
-    requiredListAttrs = ['project_id', 'snippet_id']
-    requiredGetAttrs = ['project_id', 'snippet_id']
-    requiredCreateAttrs = ['project_id', 'snippet_id', 'body']
+    requiredUrlAttrs = ['project_id', 'snippet_id']
+    requiredCreateAttrs = ['body']
 
 
 class ProjectSnippet(GitlabObject):
     _url = '/projects/%(project_id)s/snippets'
     _constructorTypes = {'author': 'User'}
-    requiredListAttrs = ['project_id']
-    requiredGetAttrs = ['project_id']
-    requiredCreateAttrs = ['project_id', 'title', 'file_name', 'code']
+    requiredUrlAttrs = ['project_id']
+    requiredCreateAttrs = ['title', 'file_name', 'code']
     optionalCreateAttrs = ['lifetime']
     shortPrintAttr = 'title'
+
 
     def Content(self, **kwargs):
         url = "/projects/%(project_id)s/snippets/%(snippet_id)s/raw" % \
@@ -992,7 +1019,8 @@ class UserProject(GitlabObject):
     canDelete = False
     canList = False
     canGet = False
-    requiredCreateAttrs = ['name', 'user_id']
+    requiredUrlAttrs = ['user_id']
+    requiredCreateAttrs = ['name']
     optionalCreateAttrs = ['default_branch', 'issues_enabled', 'wall_enabled',
                            'merge_requests_enabled', 'wiki_enabled',
                            'snippets_enabled', 'public', 'visibility_level',
@@ -1003,13 +1031,13 @@ class Project(GitlabObject):
     _url = '/projects'
     _constructorTypes = {'owner': 'User', 'namespace': 'Group'}
     canUpdate = False
-    canDelete = False
     requiredCreateAttrs = ['name']
     optionalCreateAttrs = ['default_branch', 'issues_enabled', 'wall_enabled',
                            'merge_requests_enabled', 'wiki_enabled',
                            'snippets_enabled', 'public', 'visibility_level',
-                           'namespace_id', 'description']
+                           'namespace_id', 'description', 'path', 'import_url']
     shortPrintAttr = 'path'
+
 
     def Branch(self, id=None, **kwargs):
         return ProjectBranch._getListOrObject(self.gitlab, id,
@@ -1127,10 +1155,8 @@ class Project(GitlabObject):
 class TeamMember(GitlabObject):
     _url = '/user_teams/%(team_id)s/members'
     canUpdate = False
-    requiredCreateAttrs = ['team_id', 'user_id', 'access_level']
-    requiredDeleteAttrs = ['team_id']
-    requiredGetAttrs = ['team_id']
-    requiredListAttrs = ['team_id']
+    requiredUrlAttrs = ['teamd_id']
+    requiredCreateAttrs = ['access_level']
     shortPrintAttr = 'username'
 
 
@@ -1138,11 +1164,10 @@ class TeamProject(GitlabObject):
     _url = '/user_teams/%(team_id)s/projects'
     _constructorTypes = {'owner': 'User', 'namespace': 'Group'}
     canUpdate = False
-    requiredCreateAttrs = ['team_id', 'project_id', 'greatest_access_level']
-    requiredDeleteAttrs = ['team_id', 'project_id']
-    requiredGetAttrs = ['team_id']
-    requiredListAttrs = ['team_id']
+    requiredCreateAttrs = ['greatest_access_level']
+    requiredUrlAttrs = ['team_id']
     shortPrintAttr = 'name'
+
 
 
 class Team(GitlabObject):
@@ -1150,6 +1175,7 @@ class Team(GitlabObject):
     shortPrintAttr = 'name'
     requiredCreateAttrs = ['name', 'path']
     canUpdate = False
+
 
     def Member(self, id=None, **kwargs):
         return TeamMember._getListOrObject(self.gitlab, id,
